yolo5/requirements.txt
pyyaml
loguru
# AWS integration required modules
boto3
botocore
# Polybot microservices interaction required modules
pymongo
requests